Moved some functions to ood_detect.py, Added new file for augmentations related functions
Minor changes - moving functions to the right files as suggested
